3rdParty-devel: Patch plib
There are several errors when building plib for Windows:

- Casting a pointer to DWORD is unsafe because they might have different
sizes. DWORD_PTR [1] must be used instead.
- Casting a pointer to long is unsafe because they might have different
sizes. ptrdiff_t must be used for pointer arithmetic.
- slDSP::counter is private so it cannot be used by waveOutProc. A
public function is now defined to achieve the expected behaviour.

[1]: https://learn.microsoft.com/en-us/openspecs/windows_protocols/ms-dtyp/66d61dd8-2191-4a37-b963-e49bf0dc2579


git-svn-id: https://svn.code.sf.net/p/speed-dreams/code/trunk@9579 30fe4595-0a0c-4342-8851-515496e4dcbd

Former-commit-id: d284052bb28c49e40714a36cdc135aa284a524fa
Former-commit-id: a5f73738a1f8f9a1687c5e071f20c99b0e83a4dd

+++ packaging/3rdParty-devel/patches/plib/ssgParser.cxx
@@ -36,6 +36,7 @@
 
 #include "ssgLocal.h"
 #include "ssgParser.h"
+#include <cstdint>
 
 
 static _ssgParserSpec default_spec =
@@ -263,7 +264,7 @@ void _ssgParser::expectNextToken( const char* name )
 // This is copied to a new buffer, so that I have the space to add the 0.
 void _ssgParser::addOneCharToken ( char *ptr ) 
 {
-	assert( (long)onechartokenbuf_ptr- (long)onechartokenbuf < 4096 ) ; // Buffer overflow
+	assert( (ptrdiff_t)onechartokenbuf_ptr- (ptrdiff_t)onechartokenbuf < 4096 ) ; // Buffer overflow
 	
 	onechartokenbuf_ptr [ 0 ] = *ptr;
 	onechartokenbuf_ptr [ 1 ] = 0;
